refactor form navigator into base class
